optimizations on sequences
